Initial commit
Okay, I would like to do some financial modeling with real stock data. I want to compare various profit-sharing strategies.

First I need a framework. As an exercise, let's give it a GUI. How about tkinter and python?

The program opens with a large pane with data entry fields for stock ticker, qty shares, start date, end date.

Ultimately we are going to have a data window to show the chart over that time as well as some calculations that require all the daily stock data for that stock to be retrieved for both display and use in the calculations.

Example algorithm:

Assume an equal-sized cash position. Rebalance the asset when it differs by 5% either way from 50% of the total value.

This can be contrasted with buy-and-hold algorithm, etc.

This is a completely empty folder so how about a prototype GUI with some things stubbed out for the future?

Is that enough guidance for what I want?